refactor piceces into components.tsx
